move mappings inside the coroutines
