+++ app/src/androidTest/java/com/jeff/pizza/test/UserLoggedInTest.kt
@@ -4,19 +4,19 @@ import com.jeff.pizza.R
 import com.jeff.pizza.base.BaseFragmentTest
 import com.jeff.pizza.core.data.repository.user.UserDataSourceRepository
 import com.jeff.pizza.core.domain.model.user.UserType
+import com.jeff.pizza.core.presentation.model.ProductUI
 import com.jeff.pizza.di.ApplicationModule
 import com.jeff.pizza.mockers.marriedProducts
 import com.jeff.pizza.mockers.pizzaEspencatDetails
 import com.jeff.pizza.mockers.pizzaEspencatMarriedDetails
 import com.jeff.pizza.mockers.pizzaMargaritaDetails
-import com.jeff.pizza.mockers.singleProducts
 import com.jeff.pizza.mockers.shoppingCartInfo
 import com.jeff.pizza.mockers.shoppingCartMarriedInfo
+import com.jeff.pizza.mockers.singleProducts
 import com.jeff.pizza.pageobject.ProductDetailsPageObject
 import com.jeff.pizza.pageobject.ProductsPageObject
 import com.jeff.pizza.pageobject.ShoppingCartPageObject
 import com.jeff.pizza.pageobject.SplashPageObject
-import com.jeff.pizza.products.presentation.model.ProductUI
 import com.jeff.pizza.utils.readStringFromFile
 import dagger.hilt.android.testing.HiltAndroidRule
 import dagger.hilt.android.testing.HiltAndroidTest
