+++ core/src/main/java/com/jeff/pizza/core/domain/usecase/GetProductsUseCase.kt
@@ -2,15 +2,16 @@ package com.jeff.pizza.core.domain.usecase
 
 import com.jeff.pizza.core.domain.model.base.Either
 import com.jeff.pizza.core.domain.model.base.Failure
-import com.jeff.pizza.core.domain.model.products.Product
 import com.jeff.pizza.core.domain.resource.products.ProductsResource
 import com.jeff.pizza.core.domain.resource.user.UserResource
+import com.jeff.pizza.core.presentation.model.ProductUI
+import com.jeff.pizza.core.presentation.model.toUI
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.withContext
 import javax.inject.Inject
 
 interface GetProductsUseCase {
-    suspend fun execute(refresh: Boolean): Either<Failure, List<Product>>
+    suspend fun execute(refresh: Boolean): Either<Failure, List<ProductUI>>
 }
 
 class GetProductsUseCaseImpl @Inject constructor(
@@ -18,9 +19,9 @@ class GetProductsUseCaseImpl @Inject constructor(
         private val productsResource: ProductsResource,
         private val dispatcher: CoroutineDispatcher): GetProductsUseCase {
 
-    override suspend fun execute(refresh: Boolean): Either<Failure, List<Product>> {
+    override suspend fun execute(refresh: Boolean): Either<Failure, List<ProductUI>> {
         return withContext(dispatcher) {
-            productsResource.getProducts(refresh, userResource.getUserType())
+            productsResource.getProducts(refresh, userResource.getUserType()).map { it.toUI() }
         }
     }
 }
