+++ products/src/main/java/com/jeff/pizza/products/domain/usecase/GetProductUseCase.kt
@@ -1,23 +1,23 @@
 package com.jeff.pizza.products.domain.usecase
 
-import com.jeff.pizza.core.domain.model.products.Product
 import com.jeff.pizza.core.domain.resource.products.ProductsResource
-import com.jeff.pizza.core.domain.resource.user.UserResource
+import com.jeff.pizza.products.presentation.model.ProductDetailsUI
+import com.jeff.pizza.products.presentation.model.toDetailsUI
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.withContext
 import javax.inject.Inject
 
 interface GetProductUseCase {
-    suspend fun execute(productId: Long): Product
+    suspend fun execute(productId: Long): ProductDetailsUI
 }
 
 class GetProductUseCaseImpl
 @Inject constructor(private val productsResource: ProductsResource,
                     private val dispatcher: CoroutineDispatcher): GetProductUseCase {
 
-    override suspend fun execute(productId: Long): Product {
+    override suspend fun execute(productId: Long): ProductDetailsUI {
         return withContext(dispatcher) {
-            productsResource.getProduct(productId)
+            productsResource.getProduct(productId).toDetailsUI()
         }
     }
 }
