+++ products/src/main/java/com/jeff/pizza/products/presentation/model/ProductsUIState.kt
@@ -1,5 +1,7 @@
 package com.jeff.pizza.products.presentation.model
 
+import com.jeff.pizza.core.presentation.model.ProductUI
+
 
 sealed class ProductsUIState {
     object Loading: ProductsUIState()
